Fix issues with some operators in activity report
This fixes a number of issues with the "Is not one of", "Is empty (Null)"
and "Is not empty (Null)" operators for the "Campaign(s)", "Assigned To"
and "Created By" fields. These operators previously did not work as the
generated SQL was invalid. Additionally, for the "Is not one of" operator,
the query was changed so it included values with a NULL value, which
is consistent with core behaviour.